backend/requirements.txt
# Django and DRF - Framework principal y API REST
Django==5.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==24.3

# Database - Conectores de base de datos
# PyMySQL==1.1.0         # MySQL connector alternativo (pure Python)
# mssql-django==1.4    # SQL Server support (opcional, instalar solo si se requiere)

# Development and Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.3.0
factory-boy==3.3.0

# Code Quality
black==23.7.0
# Note: isort and flake8 were removed — project uses Black as the single formatter/linter.

# Production Server
gunicorn==21.2.0
whitenoise==6.5.0

# Health Checks and Monitoring
django-health-check==3.17.0

# Security
django-ratelimit==4.1.0
python-decouple==3.8

# File handling
Pillow==11.3.0
django-storages==1.14.2

# Utilities
celery==5.3.1
redis==4.6.0

setuptools
